channels/discord_reader.py
"""
Discord channel reader implementation.
"""

import logging
from typing import List, Dict, Any
import discord

from .base_reader import BaseChannelReader

logger = logging.getLogger(__name__)

class DiscordReader(BaseChannelReader):
    """
    Reads posts from Discord channels using discord.py library.
    """

    def __init__(self, config: Dict[str, Any]):
        super().__init__(config)
        self.bot_token = config.get('DISCORD_BOT_TOKEN')
        self.channel_id = config.get('DISCORD_CHANNEL_ID')
        self.client = None

    def authenticate(self) -> bool:
        """
        Authenticate with Discord API.

        Returns:
            True if authentication successful, False otherwise
        """
        try:
            intents = discord.Intents.default()
            intents.messages = True
            intents.message_content = True

            self.client = discord.Client(intents=intents)

            @self.client.event
            async def on_ready():
                logger.info(f"Discord bot logged in as {self.client.user}")

            # Note: This is synchronous, actual authentication happens when running the client
            return True

        except Exception as e:
            logger.error(f"Discord authentication error: {str(e)}")
            return False

    def read_posts(self, limit: int = 100) -> List[Dict[str, Any]]:
        """
        Read posts from Discord channel.

        Args:
            limit: Maximum number of posts to read

        Returns:
            List of standardized post dictionaries
        """
        # Note: Discord.py requires async context, this is a simplified version
        # In a real implementation, you'd need proper async handling
        logger.warning("Discord reader not fully implemented - requires async context")
        return []
